clear raw files

+++ main.py
@@ -16,9 +16,15 @@ async def download_file(client, url):
     Downloads a file and stores it in OUTPUT_DIR.
     Returns the full filename.
     '''
+
+    # Check if file already exists
+    filename = OUTPUT_DIR + os.path.basename(url)
+    if os.path.isfile(filename):
+        return filename
+
+    # Make the request
     async with client.get(url) as response:
         if response.status == 200:
-            filename = OUTPUT_DIR + os.path.basename(url)
             f = await aiofiles.open(filename, mode='wb')
             await f.write(await response.read())
             await f.close()
@@ -48,7 +54,7 @@ def main():
         ROOT + '/data-center/student/application-volume/fafsa-school-state')
     soup = BeautifulSoup(response.content, 'lxml')
 
-    # Find the dropdown menu
+    # Find the dropdown menu, send all options into our downloader
     dropdown = soup.find(id='gourl1')
     return dict(asyncio.run(download_all_options(dropdown.find_all('option'))))
